Added files
I'm sorry, I directly open the html files to my browser to see my output. I added My name to the html and my email. And I edited the javascript to be able to take an entry and rewrite the header as well as give an alert with the given text entry